+++ stdlib/LinearAlgebra/test/tridiag.jl
@@ -553,9 +553,12 @@ end
     @test diag(S + 2I) == zero(d)
 end
 
+const BASE_TEST_PATH = joinpath(Sys.BINDIR, "..", "share", "julia", "test")
+isdefined(Main, :ImmutableArrays) || @eval Main include(joinpath($(BASE_TEST_PATH), "testhelpers", "ImmutableArrays.jl"))
+using .Main.ImmutableArrays
+
 @testset "Conversion to AbstractArray" begin
     # tests corresponding to #34995
-    using LinearAlgebra: ImmutableArray
     v1 = ImmutableArray([1, 2])
     v2 = ImmutableArray([3, 4, 5])
     v3 = ImmutableArray([6, 7])
